Refactor to remove duplication
